sql/row: split the FK span generation into a separate file
Release note: None
