+++ pkg/sql/row/fk.go
@@ -755,55 +755,3 @@ func makeBaseFKHelper(
 		)
 	}
 }
-
-func (f baseFKHelper) spanForValues(values tree.Datums) (roachpb.Span, error) {
-	var key roachpb.Key
-	if values != nil {
-		span, _, err := sqlbase.EncodePartialIndexSpan(
-			f.searchTable.TableDesc(), f.searchIdx, f.prefixLen, f.ids, values, f.searchPrefix)
-		return span, err
-	}
-	key = roachpb.Key(f.searchPrefix)
-	return roachpb.Span{Key: key, EndKey: key.PrefixEnd()}, nil
-}
-
-func (f baseFKHelper) span() roachpb.Span {
-	key := roachpb.Key(f.searchPrefix)
-	return roachpb.Span{Key: key, EndKey: key.PrefixEnd()}
-}
-
-// FkSpanCollector can collect the spans that foreign key validation will touch.
-type FkSpanCollector interface {
-	CollectSpans() roachpb.Spans
-	CollectSpansForValues(values tree.Datums) (roachpb.Spans, error)
-}
-
-var _ FkSpanCollector = fkInsertHelper{}
-var _ FkSpanCollector = fkDeleteHelper{}
-var _ FkSpanCollector = fkUpdateHelper{}
-
-func collectSpansWithFKMap(fks map[sqlbase.IndexID][]baseFKHelper) roachpb.Spans {
-	var reads roachpb.Spans
-	for idx := range fks {
-		for _, fk := range fks[idx] {
-			reads = append(reads, fk.span())
-		}
-	}
-	return reads
-}
-
-func collectSpansForValuesWithFKMap(
-	fks map[sqlbase.IndexID][]baseFKHelper, values tree.Datums,
-) (roachpb.Spans, error) {
-	var reads roachpb.Spans
-	for idx := range fks {
-		for _, fk := range fks[idx] {
-			read, err := fk.spanForValues(values)
-			if err != nil {
-				return nil, err
-			}
-			reads = append(reads, read)
-		}
-	}
-	return reads, nil
-}
